+++ src/decorators/index.ts
@@ -4,3 +4,6 @@
 export { BotDeclaration } from "./BotDeclaration";
 export { ConnectorDeclaration } from "./ConnectorDeclaration";
 export { OutgoingWebhookDeclaration } from "./OutgoingWebhookDeclaration";
+export { MessageExtensionDeclaration } from "./MessageExtensionDeclaration";
+export { PreventIframe } from "./PreventIframe";
+export { BotCallingWebhook } from "./BotCallingWebhook";