+++ src/interfaces/index.ts
@@ -1,6 +1,5 @@
 // Copyright (c) Wictor Wilén. All rights reserved.
 // Licensed under the MIT license.
 
-export { IBot } from "./IBot";
 export { IConnector } from "./IConnector";
 export { IOutgoingWebhook } from "./IOutgoingWebhook";